Attaching non-boolean non-changers to hooks now produces a runtime error.
This complements the recent relaxation of the hook attachment syntax, and patches up a silent variety of type error in Harlowe.

JS's undefined and null will now cause this error to result instead of acting like false, but since they can't be accessed from normal Harlowe, this shouldn't be too big an issue.