The CodeMirror mode's list of macro names is now generated at compile time.
The documentation comments are used to populate an object, which is injected into markup/codemirror/mode.js. This prevents the data from becoming desynchronised with the codebase. It's also possible to expand this in the future to provide other data to the mode, such as autocomplete, rudimentary type signature checking, or popup information.

Also fixed a bug where changers could wrongly attach to named (tagged) macros.