Added column markup syntax.
Like aligner markup, column markup consists of single-line tokens indicating that the subsequent text should be separated into columns. They consist of a number of | marks, indicating the size of the column relative to the other columns, and a number of = marks surrounding it, indicating the size of the column's margins in CSS "em" units (which are about the width of a capital M). Separate each column's text with tokens like |=== and ==||, and end them with a final |==| token to return to normal page layout.

Also altered the general CSS so that consecutive line breaks are now slightly shortened.