Multiple changer values can now be attached to single hooks.
This means that, for instance, "(text-style:'bold')(align:'==>')$robotFont[Text]" acts like attaching the concatenation of each of those changers to the [Text] hook, saving the need to (set:) them together into a variable.

Changers will attempt to chain up to immediately succeeding expressions, adding their values together. If the result is an error, then the chain stops there and then. If the chain never reaches an anonymous hook, then the existing error for unattached changers is used.

The intervening whitespace between the changers' expressions is, of course, deleted.

Closes #2.

Also fixed a bug with the CSS in the previous commit, and removed the prevTextNode() and nextTextNode() jQuery extensions.