Initial commit of the debug mode variable pane.
This pane displays the (currently just global) variables in the story, and should update them the moment they change, using VarRef event handlers. This still could do with a few more features, such as: a button to hide/show the pane, and of course the inclusion of temporary variables.

Also:

 * Fixed some serious bugs in the debug mode's turns dropdown.

 * Added an on() event handler register to VarRef.

 * Error messages now refer to Booleans as "the Boolean value" rather than "the logic value".

 * Corrected a few documentation lines.