Added a string-specific shorthand of (repeated:) called (str-repeated:).
This will be followed by a number of other str- macros in the coming days.

Additionally, the (text:) macro now has another alias, (str:). This alias will now be the preferred name for this macro in the documentation, mainly due to the arrival of other string macros that begin with "str-", and additionally to avoid semantic conflict with the various "text-" changer macros like (text-style:).