Added (click-goto:), (mouseover-goto:) and (mouseout-goto:), as shorthands for the (click:), (mouseover:) and (mouseout:) macros combined with (goto:).
The primary incentive for doing this was to add a special form of the link syntax, with no link text, resembling [[->passage name]], as syntactic sugar for (click-goto: ?Page, "passage name") etc., so as to promote the use of full-page clicking in stories, as an alternative, in strictly linear stories, to single links of arbitrary words. But, the appeal of this shorthand, as an alternative way to create otherwise normal links, is also desirable.