Implemented (transition-depart:) and (transition-arrive:) as a means to change the passage transitions used by links.
(t8n-depart:) and (t8n-arrive:) only have an effect when affixed to (link-goto:) and (link-undo:) (as well as when (enchant:?Link) targets those entities) and alter the between-passage transitions of those links. Fortunately, the existing transitions' code didn't need to be adjusted to work with passages. The terms "depart" and "arrive" may be a little long, but I like how unambiguously they convey the idea of travel between passages.

Also added "instant" as a valid transition name.