Finished implementing (cycling-link:) and VarBind.
I'm still not 100% confident on the "bind" keyword, but I guess this is what I'm sticking with.

You may notice that a VarBind stores a VarRef, which contains a variable name and property chain, but afterward one can then alter the variable from under it, so that a property chain for a datamap is now applied to an array. canSet() in varref.js has some provisions to make these not much of a hassle, but I changed it so that internally, VarRefs' property chains now store array indexes as numbers instead of strings, so that they might be distinguished from datamap names. This isn't ideal at all, but for now it produces user-facing errors when it should.

Working on this implementation leaves me with a few additional questions:

 * Whether (cycling-link:) rejecting an empty link text should be universalised to the other link macros.

 * Whether (t8n:) attached to a (cycling-link:) causing it to transition in on passage start is a problem. (Should it follow how (link:) interacts with t8n?)