Renamed "HookCommands" back to just Commands, and reworded the documentation for those commands that can't have changers attached.
My reasoning is that the commands that can't have attachments are generally self-evident - (enchant:), (load-game:), (goto-url:), etc. have no obvious entity or effect to be styled - so leaving them as special cases shouldn't be too onerous.

Also changed (show:) to be a HookCommand (that is, it can have changers attached).