Added a (link-show:) macro, a link which, when clicked, shows hidden hooks as if by (show:).
I'm really conflicted on whether this should be a combination of (link-reveal:) or (link-replace:) - that is, whether the link text, when clicked, should unlink or disappear. I've gone with the former for now because I suspect the writing style it would afford - links within sentences, rather than links as single "commands" - is more popular for most Twine authors.

Also, having noticed that a bug prevents (show:) from throwing errors when made to show an already visible hook, I decided to make that the intended behaviour and amend the documentation.