Temp variables finally work correctly with changers that defer a hook.
Now, you can reference temp variables inside the hook, such as in (link:"Read")[It reads: _engraving], just as you can with other kinds of changers.

Additionally:

 * Error messages should now explain in slightly more detail what kind of lambda a macro requires ("a 'where' macro", for instance).
 * The "undo" and "redo" buttons' CSS opacity has been reduced, to better fit the default dark theme.
 * Fixed a number of macros that shouldn't have been made into commands, like (prompt:) or (save-game:).

As a result of that last point, three macros that used to return { earlyExit: 1 }, (save-game:), (goto-url:) and (reload:), no longer do, because Section.runExpression no longer recognises that as a valid return value outside of attachables with TwineScript_Run() (which (goto:) is a part of). I ought to fix this at some later point, though the low traffic of those three makes this a low priority.