Lambda macros that accept multiple values no longer error if no values are given after the lambda.
The macros are (all-pass:), (some-pass:), (none-pass:), (find:), and (altered:). (folded:) is not included. The values returned by (all-pass:), (some-pass:) and (none-pass:) getting no values are in accordance with Haskell and other FP libraries for JavaScript.

Closes #145.

Also fixed a bug where having a (for:) macro's lambda's "where" clause return something other than a boolean, such as in (for: each _a where 127)[], wouldn't produce an error message.