Fixed a bug where the new SessionStorage system would interfere with the "test story starting here" feature in Twine debugging.
Now, that feature is ignored if the debug option is enabled.

Also, if the (loadgame:) macro encounters an error while loading save data (such as, a passage it refers to no longer exists in this version of the story) then a user-facing prompt() will appear suggesting that the save data might be outdated, and asking the reader whether or not the save data should be deleted. It's expected that this message should only appear during story testing, or when upgrading to later versions of a story.