Implemented a new "visits" identifier, a (passages:) macro, and changed (history:) to accept an optional filtering "where" lambda.
The purpose of providing a filtering lambda to (history:) is to make it easier to replicate the Twine 1 visitedtags() function, which returns how many times the player visited a passage with a given tag or tags. This can be accomplished by filtering the (history:) array using (find:) and checking its length, and this new feature offers a convenient shorthand for that.

The purpose of the "visits" identifier is to make it easier to replicate the Twine 1 <<once>> macro, which only displayed text on the first visit to a passage, and whose absence is a long-standing weakness in Harlowe. Now, it can be replicated using (if: visits is 1), which expresses the intent much better by leveraging the existing (if:) macro.

The new (passages:) macro is mainly to just provide a more general querying of passages than the simple search-by-name offered by (passage:).

Closes #135.