Added a new array, string and hookset slicing dataname syntax, of the form "1sttolast".
I have some misgivings about this syntax (I'd been hoping to allow hyphens/minus signs between the "to"), but I think it will provide a convenient and readable alternative to the (a:) dataname syntax for the most basic usecase, slicing on known indices.

Actually, my main motivation was to provide a syntax that would avoid the "positive-negative range" problem with just using (range:) in that position: (range:1,-1) produces a 3-element array that can't be used for indexing, whereas "1sttolast" works as intended - so, that, at least, is an improvement.

I feel like I might soon need to add one final replacement for (range:), possibly called (slice:), which is used only in the position normally occupied by (range:).