Harlowe crash messages (for both user scripts and engine crashes) now use the new HTML dialogs.
Also fixed a longstanding bug where the CSS code to make consecutive <br>s cumulatively smaller was erroneously being applied to a lot of other <br>s. The CSS has been replaced with special Renderer checks and inline styles for the <br>s.

Additionally, the colour scene of <tw-dialog>s is now white-on-black regardless of <tw-passage> CSS. Whether or not a more convenient means to change it will be implemented is currently undecided.