Added macros for a new rudimentary storylet system, (storylet:) and (open-storylets:)
This was more complex than I'd hoped, requiring me to think up a whole paradigm for executing a small portion of passage code outside of it. As it is, giving this responsibility to Renderer is a bit suspect, but I need it as close to the actual lexing and parsing as possible.

I don't know if I like the name (open-storylets:), due to "open" being ambiguous as a verb and an adjective, but it's only four letters, so it stays for now.

Also fixed a few bugs and documentation quirks.