Added a (metadata:) macro, which, when placed in a passage, adds values to the (passage:) datamap for that passage.
Once again, more contexts for external execution of passage code had to be developed for this, as well as a startup error dialog, which MIGHT be adapted to other kinds of startup errors in the future.

Also, like with (storylet:), the test code for this is required to use a different path, calling the Passage metadata setup functions manually rather than at Harlowe startup.

Also fixed a bug with the REPL debugging function.