Added a (rerun:) macro, which is similar to (show:) but re-runs already visible hooks instead of hidden hooks.
I feel like this is a prelude for me to finally add a (hide:) macro, something I'd been reticent about adding for awhile, due to how the semantics of (show:)ing a hidden hook work - whether a rerun would be performed or not. I feel like this macro is helping to turn my mind toward adding it and just letting the slight semantic inconsistency be as it may.

Also:
 * Added two new (textstyle:)s, adjusted the CSS of "rumble" and "shudder", and altered the macro to accept multiple style options to combine.
 * Changed the documentation for (subarray:) and (substring:) marking them as un-deprecated, and explaining the situations when they should be used over the normal syntax.