Added a dark mode selector, a code preview pane and a CodeMirror editor to the Harlowe docs.
This has been an aspiration of mine for awhile, so I'm glad it's done. Problem is, a lot of the code samples in the documentation need to be updated so that the variables they reference actually exist, so that they don't produce errors when run impromptu.

Also fixed a long-standing lambda bug that prevented lambda temp variables from having capital letters.