Added (hide:), a counterpart to (show:) that causes the hidden block to not re-render when shown again.
Again, I think the semantics of showing (hide:) hooks versus hidden hooks is a little uncomfortable, but I think the strength of the word "hide" would allow it to make sense to the author - and especially since the name of (rerun:) provides a contrast to it.

Closes #86.

Also added the debug mode variables pane to the documentation's live code preview, and fixed an error message bug involving computed properties of datamaps.