Upgraded (enchant:) to accept a 'via' lambda as a second argument, and allow it to add transitions to its scopes.
This is paving the way to implement the "typewriter" effect into Harlowe using only its own macros and types. All that's left is (t8n-delay:).

I acknowledge (enchant:)'s type signature now looks a little awkward, but I consider it to be special compared to other commands, and that each of its arguments' two types has some similarity with the other (for the second, changers and 'via' lambdas returning changers) - so I'll tacitly accept this.

While working on this, I fixed the bug where (t8n:)-augmented hooks' animations would restart whenever the passage had anything appended to it, and hammered out a proper semantic for (t8n:) and replacement macros. Now, hooks keep timestamps of their last modification, and any (enchant: (t8n:))s added to the passage that are newer than those timestamps are expedited by the delta (i.e. if the animation has finished, it isn't restarted).

Also refactored a bit of the transition code in Utils.