Added the (transition-delay:) macro.
This completes the functionality for the "typewriter" text effect, as mentioned previously. In order to make this work, some subtle CSS changes had to be made to the transitions, specifically giving .transition-in elements visibility:hidden, which each animation removes.

Also worked to optimise the enchantScope function in enchantments, which bogged down a lot when enchanting ?passage's chars or ?passage's lines.