Added four border-related changer macros.
These macros pose something of an issue, in that a lot of CSS border functionality (in particular, individual directional border settings) is not offered by them, but there isn't an obvious way to make most of it available without complicating the type signatures of these macros a lot, or filling the macro namespace with unsightly combinations like (border-top-width:). Currently, the intention is to simply let (css:) be the fallback, but this may change.

Also:
 * Fixed a crash involving (enchant: ?Passage's lines).
 * Changed a couple of TwineErrors' types from "macrocall" to "datatype".