Added (palette:), a rapid prototyping colour-palette generator macro, as well as various LCH colour space tools - the "lch" datamap on Colours, the (lch:) macro, and (complement:).
I needed to add LCH color space support to Harlowe because (palette:), in order to generate high-contrast text and background colours, needed to operate in a colour space that properly modeled luminosity, instead of just "lightness". The other new features came as a result of implementing (palette:).

(palette:) is not yet complete, though, because I haven't finalised its output for the latter values, or decided whether I even want the fourth value in the first place. This macro is also based on an as-yet unwritten macro, (text-palette:) - which I will probably rename when I do write it - and it's important that this macro's output and that macro's input are aligned.