Implemented the major functionality of custom macros, as well as their related datatypes.
The CustomMacro data type has been added, which handles most of the behaviour of custom macro execution.

This also changes some internals in Macros to accommodate custom macros. TypeSignatureCheck is now a function that's called by Macros.run() rather than a wrapper that gets created for every built-in macro.

Three important areas remain to be written: error-propagation to extract errors from inside custom macros' code hooks and return them to the macro itself, tests need to be written, and documentation needs to be completed.