Added the ability to type-restrict variables using TypedVars, (set:) and (put:).
This required adding a new structure to variables stores, TypeDefs, which holds the type definitions. I'd wanted this feature to cover deep structures in datamaps, but just having this feature available for basic variables is decent enough.

Custom macros' arguments are now type-restricted, as well.

Also added the "lambda", "macro" and "bool" datatypes (one being an alias of "boolean").