Finally added (input-box:) and, at the spur of a whim, (force-input-box:).
These macros' twisted and gnarled type-signatures irk me (and largely precluded me adding this macro earlier), but I don't know another way around the constraints that they be consistent with both (cycling-link:) and (box:)'s parameter orderings. Also, I was forced to limit the sizing line strings given to this and (box:), where a one-character "X" is the only valid 100%-width string, to avoid ambiguity with the other string that this takes. Regrettable.

Also fixed a few more minor bugs.

Closes #130. Closes #132.