Added the "random" dataname to arrays and strings.
For awhile I've wanted a "deck" datatype that lets you draw random values from it while also potentially adding values to it, which isn't as convenient with just (shuffled:) - but this, combined with (move:), should allow unshuffled arrays to serve that purpose. Note that (shuffled:) is still useful if you want to reuse the same random sequence repeatedly.

Also:
 * Added a "fidget" text style, inspired by Undertale, as well as variations of "underline" and "strike", and fixed several styles being incompatible with each other.
 * Now, propagated errors from custom macros are no longer listed in the error log.
 * Now, trying to access the "0th" or "0thlast" value in an array or string produces an error.
 * Fixed a bug where "is not an" would cause an error.