Began implementing destructuring assignment of array and datamap contents.
This has proven to be a much more awkward task than I'd hoped... I still have to nail down a couple of details, such as whether the actual (a:) and (dm:) macros should be used at all for patterns, given that A) the semantics of destructuring matching are looser than that of the exact matching of "matches", B) this implies that TypedVars are storable values now, which they really should not be, and C) my plan for string pattern objects requires said patterns to be datatype data rather than actual arrays.

Also added the 'integer', 'lowercase', 'uppercase', 'newline' and 'alphanumeric' datatypes, and added some bug report writing suggestions (formerly in the Bitbucket repository data) to the manual.