Implemented destructuring patterns that use untyped variables, as well as expanding TypedVars to include any pattern as the 'type' instead of just datatypes.
This brings destructuring syntax roughly in line with the fluidity of JS destructuring. This was implemented by altering compile() to compile all variables
in the left-hand-side of a 'to' or 'into' statement into an 'any-type' VarRef.

Also fixed some bugs in the previous commit, and made the alpha of <tw-error> elements denser.