Began implementing spread datatypes, which represent zero or more of their values.
This replaces "spread typedvars" with a more general concept. The most bothersome part of this - array destructuring compatibility - has yet to be implemented, however.

In order to make this work, the specificity of ... needed to be changed. Hopefully, as it's an otherwise limited operator, this won't have any deleterious effects.

Also added (datatype:) and (datapattern:).