Added (click-undo:) and its ilk, to complement (click-goto:).
This was actually added so that The Harlowe Toolbar would have a slightly easier time generating correct code for "click the page to undo the turn".

Also monkeyed with the syntax highlighter so that colour literals have a stripe of their colour beneath them (represented with a gradient background).

Also fixed another bug in the previous commit's fix (for harlowedocs).