Added (enchant-in:), a changer variation of (enchant:) which limits the enchantment to just the attached hook.
I don't like this as the only source of restricted-effect enchantments/changes/revisions, mainly because it seems to illuminate that the hook selection syntax (?hook + ?hook, ?hook's 1st, etc.) needs to be extended to have the same flexibility as CSS selectors, at least. However, that will have to be left unconsidered for now, due to the weight of more pressing issues.

I also am tempted to rename this to (enchant-here:), simply because the "in" doesn't have very smooth readability along with the arguments. But, "here" might not really suggest that it's a changer instead of a command, and the incompleteness of "in" does help with that...

Also:
 * Fixed (source:)'s interaction with (text-colour:).
 * Added the "fade" (transition:), as an alias for "dissolve".
 * Added "New" badges to macros new to 3.2.0 in the documentation.
 * Worked on a few more tooltips for the syntax highlighter.