Reimplemented the link changers, so that only the hook is affected by changers combined with it.
For a long time, the interaction between (link:) and other changers combined with it has been terribly inconsistent, with some changers applying to just the link, others to just the revealed hook, and most of them applying to both. I've decided after much heart-rending that I need to change this to a single hard-and-fast rule: changers combined with (link:) will only apply to the revealed hook, NOT the link. This is to allow (t8n:), (box:), (live:) and other hook-specific changers to interact with (link:) as expected, and also to make (link:) links more consistent with (click:) links.

Sadly, of course, this means that (hover-style:), whose example usage involved combining it with (link:), has now had to have its intended uses rewritten. Also, the current means of applying changers to just the link is now limited to just (enchant:) and (enchant-in:), both of which are somewhat cumbersome. Possibly an (enchant-link:) or (enchant-link-in:) macro may need to be added to make this easier.

Also fixed a bug where (for:) couldn't be attached to commands.