Added (split:) and (joined:), two basic string macros.
The reason I'd been reticent to add (split:) for so long is that I needed to see exactly what string patterns would resemble before committing to this macro's type signature. Probably overprudent in retrospect, given how straightforward it ultimately is, but nevertheless, that is how it went. That being said, I also wonder if I should a "via" lambda to (joined:)'s signature as an alternative first parameter.

Also:

 * Fixed a bug where (click:) and its ilk would cause a crash when given to (enchant:), similar to (link:) formerly.
 * Corrected some large factual errors in (folded:)'s documentation, as well as added some clarification.

Closes #15.