Renamed (output:) and (output-hook:) to (output-data:) and (output:).
I decided that having the "hook" output be the default case is more in keeping with custom macros being proposed as a more useful alternative to (display:), and that creating a lot of data macros is a more advanced case, suited to people programming more sophisticated micro-engines in Harlowe.

Also did some tweaks to the tooltips and The Harlowe Toolbar.