Added (link-fullscreen:), a macro to create a fullscreen mode toggle link, and a fullscreen sidebar button.
I've done some thinking about fullscreen functionality lately, and I've decided that there isn't THAT much reason to give the author a good degree of control over the fullscreen state. Not only is it true that most browsers nowadays only allow fullscreen mode to be entered in rather specific circumstances, but, from a UX perspective, fullscreen mode should just be a user display option, and not a necessary narrative device that must be enabled to experience the story, and which can be enabled and disabled arbitrarily - which is how it is in most all desktop games. So, I've given up plans to add fullscreen utilities more precise that these.

I don't feel very comfortable about the fullscreen button, and am likely to change it or make it optional/conditionally appendable (by the addition of explicit "icon" command macros that replicate sidebar buttons).

Also made the syntax highlighter tooltips only appear after 90 milliseconds, so as to keep from cluttering the view as you type.

Closes #35.