Added (link-style:), (line-style:) and (char-style:) as shorthands for using (enchant-in:) with ?Link, ?page's lines, and ?page's chars.
I'd lately begun to feel that it was pretty silly that such useful functionality for line and char styling was locked behind such awkwardly esoteric devices as these hook data names. More relevantly, the arrival of (enchant-in:) resulted in some messy idioms resulting from it (You can't say "this hook's lines", for instance, and instead must specify the equivalent of "the page's lines which are in this hook") and those dearly needed some alternative.

Also:
 * Fixed a bug where a "lines" enchantment would often cause text nodes within lines, such as the text of a (link:) link, to be transplanted out of any elements containing them.
 * Fixed a bug in the syntax highlighter (introduced a few commits ago) where whitespace at the start of a line would be incorrectly highlighted as an error even if it was inside a macro call.