After much anguish, an array input interface was added to the Var panel of The Harlowe Toolbar.
The big problem with this was that it involved the ability to create and remove any number of subrows (array elements) in which data could be placed, and how this interacted with the existing panel row infrastructure. That and styling, of course. But, I seem to have it in a usable spot at present.

Also, now, (hook:) works correctly when given to (enchant:)