Added (animate:), a command to re-apply a transition to an already existing hook.
This was added simply because I noticed people could use (t8n:) and (rerun:) to accomplish this, and I wanted this little hack to have a proper idiom. As it is, it isn't too useful due to
the current suite of transitions involve the hook slowly appearing (with "rumble" and "shudder" notable exceptions), but maybe a few more could be added.

As it is, I added one more transition: "blur", which uses the new CSS filters to make a hook appear from a blur.

Also fixed Utils.childrenProbablyInline() a bit more.