Rendered hooks now store the lexed tokens of their children (instead of just disposing of them) for use when rendering and re-rendering.
This is one of the additional boons I was hoping to add from the Renderer rewrite added in the previous commit. It should provide a slight perf bump to (live:) macros and other such macros.

Also added a backup try-catch that returns a TwineError message if Runtime somehow does throw an error now, although doing so is now strictly a Harlowe bug.