Improved objectName(), altering both error messages and the Variables panel's listing.
This improvement provides less vague descriptions of data structures. Arrays of various elements are described as "an array with" followed by the objectName() of items, until the string description becomes too long, whereupon it terminates with "and X more items". Datamaps and datasets have a similar description.

While this has been needed for a long time, there's still several types which could do with this improvement... mainly changers.

 * The nav column is now visible, inline, at the very top of the documentation when the window size is < 1200px.
 * Also, the night mode bar is positioned better when the window size is < 1200px.
 * Fixed some bugs in the previous commit.

Closes #226.