Fixed two bugs relating to enchantments.
 * Fixed a bug where using (append:?Link) (or one of its relatives) to append text to a hook enchanted with (click:) wouldn't work correctly (the text would be placed next to the link instead of inside).
 * Fixed a bug where (enchant:), (click:), and other enchantment macros could enchant empty hooks such as |A>[] (wrapping them with <tw-enchantment> elements), even though Harlowe usually considers empty hooks to be nonexistent, and hides them with its default CSS.

For awhile I thought that second one might be a compatibility anchor, requiring this fix to be punted to 4.0... but thinking about it, the circumstances are a bit too esoteric (involving all of ?Link, (click:), and empty named hooks) to reasonably worry about. Nevertheless, a note about this is in the Readme.