Deprecated (mouseover:), (mouseout:), and all other macros beginning with "mouse", and added the (action:) link changer to replace them.
Ever since I added the optional second value to the (click:) macros, it's become clear that creating another changer for these interaction types is by far the preferable solution compared to having dozens of (mouseover:) variants in the macro namespace. Plus, being able to make the mouseover and mouseout interactions available to all kinds of links is a big win.

There is one big discomfort in how I implemented this, however: CSS/DOM structure compatibility with Harlowe 3.2 forced me to make this much more convoluted than I'd like, with multiple selectors and the "enchantment-" CSS class prefix. I'll fix all of this up in Harlowe 4.0, however. Assuming I live long enough to make it?

While I recognise that (click:?A,(action:'mouseover')) kind of reads awkwardly (is it a click or a mouseover?), this seems to be to be a flaw in the naming of (click:). In Harlowe 4.0, I might rename it to (enlink:).

I also recognise that this might be confused with SugarCube's <<actions>> macro... but I still think this is the best name for this macro. Not just because names like (link-style:) and (event:) have already been taken.

Also:

 * Now, instead of (enchant:) checking changers' ability to be enchanted using their summary(), a canEnchant boolean is (internally) attached to them.
 * Added a string variable serialisation optimisation: if the string is an appended or prepended version of an earlier turn's string, it is serialised as a {via} valueRef accordingly.
 * Fixed the z-index layering of the preview panel, debug panel and preview code area in the docs, again.